from libertyernie/brawltools@2ca241b00c3777e0c9419e284438df760c7e015c

+++ BrawlLib/SSBB/Types/Common.cs
@@ -351,7 +351,7 @@ namespace BrawlLib.SSBB.Types
 
         public void Set(int index, ruint.RefType refType, byte dataType, VoidPtr address)
         {
-            *((ruint*) Address + index) = new ruint(refType, dataType, (int) address - Address);
+            *((ruint*) Address + index) = new ruint(refType, dataType, address - Address);
         }
     }
